webrtc/RTCPeerConnection-getIdentityAssertion.html

<!doctype html>
<meta charset=utf-8>
<title>RTCPeerConnection.prototype.getIdentityAsssertion</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="identity-helper.js"></script>
<script>
  'use strict';

  // Test is based on the following editor draft:
  // https://w3c.github.io/webrtc-pc/archives/20170605/webrtc.html

  /*
    9.6.  RTCPeerConnection Interface Extensions
      partial interface RTCPeerConnection {
        void               setIdentityProvider(DOMString provider,
                                               optional RTCIdentityProviderOptions options);
        Promise<DOMString> getIdentityAssertion();
      };

      dictionary RTCIdentityProviderOptions {
        DOMString protocol = "default";
        DOMString usernameHint;
        DOMString peerIdentity;
      };
   */

  const host = window.location.host;
  const domain = window.location.hostname;

  const alice = `alice@${domain}`;
  const bob = `bob@${domain}`;

  promise_test(() => {
    const pc = new RTCPeerConnection();
    pc.setIdentityProvider(host, {
      protocol: 'idp-test.js?foo=bar',
      usernameHint: alice,
      peerIdentity: bob
    });

    return pc.getIdentityAssertion()
    .then(assertionResultStr => {
      const { idp, assertion } = parseAssertionResult(assertionResultStr);

      const {
        watermark,
        location,
        testOptions,
        options, contents, origin,
      } = assertion;

      assert_equals(watermark, 'asserted by idp-test.js',
        'Expect watermark to be left by idp-test.js');

      assert_equals(origin, window.origin);
      assert_equals(location, `https://${host}/.well-known/idp-proxy/idp-test.js`);

      assert_equals(options.protocol, 'idp-test.js?foo=bar');
      assert_equals(options.usernameHint, alice);
      assert_equals(options.peerIdentity, bob);
      assert_equals(testOptions.foo, 'bar');

      return pc.createOffer()
      .then(offer => {
        assert_has_identity_assertion(offer.sdp, assertionResultStr);
      });
    });
  }, 'getIdentityAssertion() should load identity proxy and return assertion generated');

  /*
    RTCIdentityProviderOptions Members
      peerIdentity
        The identity of the peer. For identity providers that bind their assertions to a
        particular pair of communication peers, this allows them to generate an assertion
        that includes both local and remote identities. If this value is omitted, but a
        value is provided for the peerIdentity member of RTCConfiguration, the value from
        RTCConfiguration is used.
  */
  promise_test(() => {
    const pc = new RTCPeerConnection({
      peerIdentity: bob
    });

    pc.setIdentityProvider(host, {
      protocol: 'idp-test.js'
    });

    return pc.getIdentityAssertion()
    .then(assertionResultStr => {
      const { assertion } = parseAssertionResult(assertionResultStr);
      assert_equals(assertion.options.peerIdentity, bob);
    });
  }, 'setIdentityProvider() with no peerIdentity provided should use peerIdentity value from getConfiguration()');


</script>
